Move rest_request_before_callbacks filter to Platform_Checkout_Session
