+++ includes/platform-checkout/class-platform-checkout-session.php
@@ -10,6 +10,7 @@ namespace WCPay\Platform_Checkout;
 use Automattic\Jetpack\Connection\Rest_Authentication;
 use Automattic\WooCommerce\StoreApi\Utilities\JsonWebToken;
 use WCPay\Logger;
+use WP_REST_Request;
 
 /**
  * Class responsible for handling platform checkout sessions.
@@ -25,6 +26,33 @@ class Platform_Checkout_Session {
 	 */
 	public static function init() {
 		add_filter( 'determine_current_user', [ __CLASS__, 'determine_current_user_for_platform_checkout' ], 20 );
+		add_filter( 'rest_request_before_callbacks', [ __CLASS__, 'add_platform_checkout_store_api_session_handler' ], 10, 3 );
+	}
+
+	/**
+	 * This filter is used to add a custom session handler before processing Store API request callbacks.
+	 * This is only necessary because the Store API SessionHandler currently doesn't provide an `init_session_cookie` method.
+	 *
+	 * @param mixed           $response The response object.
+	 * @param mixed           $handler The handler used for the response.
+	 * @param WP_REST_Request $request The request used to generate the response.
+	 *
+	 * @return mixed
+	 */
+	public static function add_platform_checkout_store_api_session_handler( $response, $handler, WP_REST_Request $request ) {
+		$cart_token = $request->get_header( 'Cart-Token' );
+
+		if ( $cart_token && 0 === strpos( $request->get_route(), '/wc/store/v1/' ) && \Automattic\WooCommerce\StoreApi\Utilities\JsonWebToken::validate( $cart_token, '@' . wp_salt() ) ) {
+			add_filter(
+				'woocommerce_session_handler',
+				function ( $session_handler ) {
+					return \WCPay\Platform_Checkout\SessionHandler::class;
+				},
+				20
+			);
+		}
+
+		return $response;
 	}
 
 	/**
@@ -73,7 +101,7 @@ class Platform_Checkout_Session {
 				return null;
 			}
 
-			$session_handler = new \Automattic\WooCommerce\StoreApi\SessionHandler();
+			$session_handler = new \WCPay\Platform_Checkout\SessionHandler();
 			$session_data    = $session_handler->get_session( $payload->user_id );
 			$customer        = maybe_unserialize( $session_data['customer'] );
